refactor: 관련있는 컴포넌트끼리 폴더로 묶어준다
