+++ src/components/Player/index.js
